+++ src/router.ts
@@ -8,9 +8,14 @@ export default new Router({
   base: process.env.BASE_URL,
   routes: [
     {
-      path: '/about',
-      name: 'about',
-      component: () => import('./views/About.vue'),
+      path: '/',
+      name: 'home',
+      component: () => import('./views/Home.vue'),
+    },
+    {
+      path: '/camera',
+      name: 'camera',
+      component: () => import('./views/Camera.vue'),
     },
   ],
 });
