--- src/views/About.vue
@@ -1,6 +1,0 @@
-<template>
-  <div class="about">
-    <img alt="Yoda" src="../assets/arara.svg"/>
-    <h1>ArBC</h1>
-  </div>
-</template>
