+++ src/views/Home.vue
@@ -1,0 +1,7 @@
+<template>
+  <div class="home">
+    <h1>ArBC</h1>
+    <h1><a href="https://jlucassr.github.io/ArBC-Pages/">Sobre</a></h1>
+    <img alt="Arara" src="../assets/arara.svg"/>
+  </div>
+</template>
